Added compare and show fields options.
Now only the specified fields are used for comparison.
This is usefule to check if some modifications yield the same
solution, but not the same number of conflicts, decions, etc.

When the compared fields match the expected result. Instead of
only printing "EQUAL", now the specified "show fields" are
printed below.